build: NetBSD fixes for use of the macro toupper() and tolower()
gcc on NetBSD fails when passing char* to the to{upper,lower} macros.
Cast to a int to work around this.

CR: dalmeida@likewise.com

(lsass: r39716)

+++ lsass/common/utils/lsahash.c
@@ -466,7 +466,7 @@ LsaHashCaselessStringHash(
         // rotate result to the left 3 bits with wrap around
         result = (result << 3) | (result >> (sizeof(size_t)*8 - 3));
 
-        lowerChar = tolower(*pos);
+        lowerChar = tolower((int)*pos);
         result += lowerChar;
         pos++;
     }
